Removed the unused FileUpload.java file.
Upldated the DataCollector.

Added some global variables used for naming files. Added a copy function for copying the file to send.

Made the DataCollector write the text file header to every file it creates.

Made the DataCollector search for a gps lock if the lastLocation == bestLocation in terms of location and not timestamp (problems with network locator returning different timestamps with same location)

Updated the signal strength range.